replaced: non-frontend color use with internal datatype
